!refactor: change mysql int to bigint (#19)
* chore: add 'public' to gitignore

* feat: copied from simple-demo, basic api function

* ci: remove lint's and test's pull_request_target

* ci: lint review test

* fix: enlarge max_body_size

* style: ci fix typo

* docs: add mock test demo link

* ci: add test ci, add description for prettier

* ci: add test ci, add description for prettier

* docs: add develop guide

* revert: remove apitest

* feat: web userservice rpc client

* ci: remove gomoddirectives

* !refactor: change mysql int to bigint

+++ pkg/model.go
@@ -5,13 +5,13 @@ import (
 )
 
 type Video struct {
-	ID            int       `gorm:"column:id;primary_key;AUTO_INCREMENT"`
-	AuthorID      int       `gorm:"column:author_id;NOT NULL"`
+	ID            int64     `gorm:"column:id;primary_key;AUTO_INCREMENT"`
+	AuthorID      int64     `gorm:"column:author_id;NOT NULL"`
 	PublishTime   time.Time `gorm:"column:publish_time;default:CURRENT_TIMESTAMP;NOT NULL"`
 	FilePath      string    `gorm:"column:file_path;NOT NULL"`
 	CoverPath     string    `gorm:"column:cover_path;NOT NULL"`
-	FavoriteCount int       `gorm:"column:favorite_count;default:0"`
-	CommentCount  int       `gorm:"column:comment_count;default:0"`
+	FavoriteCount int64     `gorm:"column:favorite_count;default:0"`
+	CommentCount  int64     `gorm:"column:comment_count;default:0"`
 	Title         string    `gorm:"column:title;NOT NULL"`
 }
 
@@ -20,13 +20,13 @@ func (m *Video) TableName() string {
 }
 
 type User struct {
-	ID            int       `gorm:"column:id;primary_key;AUTO_INCREMENT"`
+	ID            int64     `gorm:"column:id;primary_key;AUTO_INCREMENT"`
 	CreateTime    time.Time `gorm:"column:create_time;default:CURRENT_TIMESTAMP;NOT NULL"`
 	Name          string    `gorm:"column:name;unique;NOT NULL"`
 	Password      string    `gorm:"column:password;NOT NULL"`
 	Salt          string    `gorm:"column:salt;NOT NULL"`
-	FollowCount   int       `gorm:"column:follow_count;default:0;NOT NULL"`
-	FollowerCount int       `gorm:"column:follower_count;default:0;NOT NULL"`
+	FollowCount   int64     `gorm:"column:follow_count;default:0;NOT NULL"`
+	FollowerCount int64     `gorm:"column:follower_count;default:0;NOT NULL"`
 }
 
 func (m *User) TableName() string {
@@ -34,10 +34,10 @@ func (m *User) TableName() string {
 }
 
 type Follow struct {
-	ID         int       `gorm:"column:id;primary_key;AUTO_INCREMENT"`
+	ID         int64     `gorm:"column:id;primary_key;AUTO_INCREMENT"`
 	FollowTime time.Time `gorm:"column:follow_time;default:CURRENT_TIMESTAMP;NOT NULL"`
-	FromUserID int       `gorm:"column:from_user_id;NOT NULL"`
-	ToUserID   int       `gorm:"column:to_user_id;NOT NULL"`
+	FromUserID int64     `gorm:"column:from_user_id;NOT NULL"`
+	ToUserID   int64     `gorm:"column:to_user_id;NOT NULL"`
 }
 
 func (m *Follow) TableName() string {
@@ -45,10 +45,10 @@ func (m *Follow) TableName() string {
 }
 
 type Comment struct {
-	ID          int       `gorm:"column:id;primary_key;AUTO_INCREMENT"`
+	ID          int64     `gorm:"column:id;primary_key;AUTO_INCREMENT"`
 	CommentTime time.Time `gorm:"column:comment_time;default:CURRENT_TIMESTAMP;NOT NULL"`
-	UserID      int       `gorm:"column:user_id;NOT NULL"`
-	VideoID     int       `gorm:"column:video_id;NOT NULL"`
+	UserID      int64     `gorm:"column:user_id;NOT NULL"`
+	VideoID     int64     `gorm:"column:video_id;NOT NULL"`
 	Content     string    `gorm:"column:content;NOT NULL"`
 }
 
@@ -57,10 +57,10 @@ func (m *Comment) TableName() string {
 }
 
 type Message struct {
-	ID          int       `gorm:"column:id;primary_key;AUTO_INCREMENT"`
+	ID          int64     `gorm:"column:id;primary_key;AUTO_INCREMENT"`
 	MessageTime time.Time `gorm:"column:message_time;default:CURRENT_TIMESTAMP;NOT NULL"`
-	FromUserID  int       `gorm:"column:from_user_id;NOT NULL"`
-	ToUserID    int       `gorm:"column:to_user_id;NOT NULL"`
+	FromUserID  int64     `gorm:"column:from_user_id;NOT NULL"`
+	ToUserID    int64     `gorm:"column:to_user_id;NOT NULL"`
 	Content     string    `gorm:"column:content;NOT NULL"`
 }
 
@@ -69,10 +69,10 @@ func (m *Message) TableName() string {
 }
 
 type Like struct {
-	ID       int       `gorm:"column:id;primary_key;AUTO_INCREMENT"`
+	ID       int64     `gorm:"column:id;primary_key;AUTO_INCREMENT"`
 	LikeTime time.Time `gorm:"column:like_time;default:CURRENT_TIMESTAMP;NOT NULL"`
-	UserID   int       `gorm:"column:user_id;NOT NULL"`
-	VideoID  int       `gorm:"column:video_id;NOT NULL"`
+	UserID   int64     `gorm:"column:user_id;NOT NULL"`
+	VideoID  int64     `gorm:"column:video_id;NOT NULL"`
 }
 
 func (m *Like) TableName() string {
